requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
bcrypt>=3.2.0
PyMySQL>=1.0.0
qrcode[pil]>=7.4.2
pytest>=8.0.0
python-dotenv>=0.19.1
Flask-SQLAlchemy>=2.5.1
Flask-Login>=0.5.0
Flask-WTF>=0.15.1
email-validator>=1.1.3
pytest-cov>=2.12.1
argon2-cffi>=21.0.0
jwt>=1.3.1
Pillow>=10.4.0
opencv-python>=4.11.0.86
mysql-connector-python>=8.0.35
